feat: new Toast API design
