+++ packages/design-system-react-native/src/components/Toast/Toast.types.ts
@@ -141,9 +141,3 @@ export type ToastProps = {
   labelsContainerProps?: Omit<ViewProps, 'children' | 'style'>;
 } & Omit<ViewProps, 'style'>;
 
-/**
- * Toast context parameters.
- */
-export type ToastContextParams = {
-  toastRef: React.RefObject<ToastRef> | undefined;
-};
