+++ packages/design-system-react-native/src/components/Toast/index.ts
@@ -4,13 +4,11 @@ export type {
   ToastOptions,
   ToastProps,
   ToastRef,
-  ToastContextParams,
   ToastLabelOptions,
   ToastDescriptionOptions,
   ToastLinkButtonOptions,
   ToastCloseButtonOptions,
 } from './Toast.types';
-export { ToastContext, ToastContextWrapper } from './Toast.context';
 export {
   TOAST_VISIBILITY_DURATION,
   TOAST_ANIMATION_DURATION,
